Finally, slim down all headerbars!
The sizing box is the culprit for the persistent chonk;
a box with a title and subtitle that is not part of the
headerbar, but is used to decide the size of the box
containing the title/subtitle. This is why headerbars
with has_subtitle=false could be shrunk and others
could not. Thanks to borschty from freenode.