gtk-css: Big restructure
This is mainly to prepare for the fact that GTK4 is
on the horizon and there will be apps using it. And
also to prepare for the fact that I will probably
have to stop using Adapta, because development on it
stopped before GTK4 was fully completed.

There is also documentation to explain the raison
d'etre and how to use the individual customizations.