scripts/depends: A script to check script dependencies.
Prints a message and sends a notification when dependencies are
missing, and exits 1, which will halt scripts. The notification is
handy in case you're just letting things go automatically but you
switch distros or something and the deps aren't there.

Of course, also migrate a bunch of scripts to use this.