conky: Might as well commit some changes I tried for improving music.
It'll be a cold day in hell before Conky gets the ability to use
fallback fonts. The issue in the tracker has languished for a long time.
So I decided I would try rendering text to an image myself. But it's
kind of a pain and my workflow was changing, so I decided to just not
use Conky anymore.