scripts/auto-steam: Specifically track only the last reaper for launch.
Another reaper starts up with the arg Install=1 when the dialog that
says "preparing to launch whatever" appears. That reaper disappears and
is then replaced by one without that arg when the game is actually
running.

We need a more reliable method to track the state of running a game. I
bet this will break in an update too.