(Sigh.) It's time to show my shameful fall. Add scripts/lib.sh.
It's exactly what you think it is. Source it and it adds a bunch of
functions and variables. I had a bunch of my scripts use it, and then
I realized I really shouldn't do that if I want these scripts to be
usable as single units. The main thing I wasn't able to pull it out of
was wino because wino is complex. ...Which means wino should become a
full program.